% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCancerStudyMetafile.R
\name{createCancerStudyMetafile}
\alias{createCancerStudyMetafile}
\title{Create Cancer Study Meta File}
\usage{
createCancerStudyMetafile(
  study_folder,
  type_of_cancer,
  cancer_study_identifier,
  name,
  description,
  citation = NULL,
  pmid = NULL,
  groups = "PUBLIC",
  add_global_case_list = NULL,
  tags_file = NULL,
  reference_genome = NULL
)
}
\arguments{
\item{study_folder}{The study folder.}

\item{type_of_cancer}{The cancer type abbreviation. This should be the same cancer type as specified in the meta_cancer_type.txt file, if available. The type can be "mixed" for studies with multiple cancer types.}

\item{cancer_study_identifier}{A string used to uniquely identify this cancer study within the database.}

\item{name}{The name of the cancer study.}

\item{description}{A description of the cancer study. This description may contain one or more URLs to relevant information.}

\item{citation}{(Optional) A relevant citation.}

\item{pmid}{(Optional) One or more relevant pubmed ids (comma separated without whitespace). If used, the field citation has to be filled, too.}

\item{groups}{(Optional) When using an authenticating cBioPortal, lists the user-groups that are allowed access to this study. Multiple groups are separated with a semicolon ";". The study will be invisible to users not in at least one of the listed groups, as if it wasn't loaded at all. See \href{https://docs.cbioportal.org/deployment/authorization-and-authentication/user-authorization/}{User-Authorization} for more information on groups. Default to 'PUBLIC'}

\item{add_global_case_list}{(Optional) Set to 'true' if you would like the "All samples" case list to be generated automatically for you. See also \href{https://docs.cbioportal.org/file-formats/#case-lists}{Case list}}

\item{tags_file}{(Optional) The file name containing custom study tags for the \href{https://docs.cbioportal.org/file-formats/#study-tags-file}{study tags}.}

\item{reference_genome}{(Optional) The study reference genome (e.g. hg19, hg38). Without specifying this property, the study will be assigned to the reference genome specified in application.properties (property ucsc.build).}
}
\description{
Create Cancer Study Meta File
}
\examples{
\dontrun{
createCancerStudyMetafile(
  study_folder = tempdir(),
  cancer_study_identifier = "brca_joneslab_2013",
  name = "Breast Cancer (Jones Lab 2013)",
  description = "Comprehensive profiling of 103 breast cancer samples.",
  add_global_case_list = TRUE
)
}
}
