% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createCancerTypeFiles.R
\name{createCancerTypeFiles}
\alias{createCancerTypeFiles}
\title{Create Cancer Type Meta File and Data File}
\usage{
createCancerTypeFiles(
  study_folder,
  data_filename = "cancer_type.txt",
  type_of_cancer,
  name,
  dedicated_color,
  parent_type_of_cancer
)
}
\arguments{
\item{study_folder}{The study folder.}

\item{data_filename}{The datafile name. If not specified, default is 'cancer_type.txt'}

\item{type_of_cancer}{The cancer type abbreviation.}

\item{name}{The name of the cancer type.}

\item{dedicated_color}{CSS color name of the color associated with this cancer study. See \href{https://www.w3.org/TR/css3-color/#svg-color}{this list} for supported names, and follow the \href{https://en.wikipedia.org/wiki/List_of_awareness_ribbons}{awareness ribbons} color schema. This color is associated with the cancer study on various web pages within the cBioPortal.}

\item{parent_type_of_cancer}{The type_of_cancer field of the cancer type of which this is a subtype. ℹ️ : you can set parent to tissue, which is the reserved word to place the given cancer type at "root" level in the "studies oncotree" that will be generated in the homepage (aka query page) of the portal.}
}
\description{
Create Cancer Type Meta File and Data File
}
\examples{
\dontrun{
createCancerTypeFiles(
  study_folder = tempdir(),
  type_of_cancer = "brca",
  name = "Breast Invasive Carcinoma",
  dedicated_color = "HotPink",
  parent_type_of_cancer = "Breast"
)
}
}
