% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{createClinicalDataFiles}
\alias{createClinicalDataFiles}
\title{Create Clinical Data Meta file and Data file}
\usage{
createClinicalDataFiles(
  study_folder,
  cancer_study_identifier,
  datatype = c("PATIENT_ATTRIBUTES", "SAMPLE_ATTRIBUTES"),
  clinical_data,
  data_filename = NULL,
  mapping_file = NULL
)
}
\arguments{
\item{study_folder}{The study folder.}

\item{cancer_study_identifier}{A string used to uniquely identify this cancer study within the database. Must be the same value specified in meta_study.txt}

\item{datatype}{'PATIENT_ATTRIBUTES' or 'SAMPLE_ATTRIBUTES'}

\item{clinical_data}{A dataframe containing the clinical data}

\item{data_filename}{The datafile name. Defaults to 'data_clinical_patient.txt' or 'data_clinical_sample.txt', depending on the datatype.}

\item{mapping_file}{A JSON file that provides field-level mapping rules to convert source data into a structure compatible with cBioPortal.}
}
\description{
Create Clinical Data Meta file and Data file
}
\examples{
\dontrun{
# Return data from REDCap
redcap_data <- redcapToCbio(
  redcap_uri = "https://redcap-dev-2.ouhsc.edu/redcap/api/"
  token = "9A068C425B1341D69E83064A2D273A70"
)

# Create clinical data files
createClinicalDataFiles(
  study_folder = tempdir(),
  cancer_study_identifier = "brca_joneslab_2013",
  datatype = "PATIENT_ATTRIBUTES",
  clinical_data = redcap_data
)
}
}
