% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStudyFolder.R
\name{createStudyFolder}
\alias{createStudyFolder}
\title{Create study folder}
\usage{
createStudyFolder(study_folder, output_path = getwd())
}
\arguments{
\item{study_folder}{The folder name.}

\item{output_path}{The path where the folder will be created.}
}
\value{
The folder path.
}
\description{
Create study folder
}
\examples{
output_folder <- createStudyFolder(study_folder = "brca")
}
