% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{.redcapToCbio}
\alias{.redcapToCbio}
\title{Convert REDCap data to cBioPortal format}
\usage{
.redcapToCbio(
  datatype = c("PATIENT_ATTRIBUTES", "SAMPLE_ATTRIBUTES"),
  redcap_uri,
  token,
  redcap_cbio_mapping,
  redcap_template = NULL
)
}
\arguments{
\item{datatype}{'PATIENT_ATTRIBUTES' or 'SAMPLE_ATTRIBUTES'}

\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{redcap_cbio_mapping}{A list that contains field-level mapping rules to convert REDCap data into a structure compatible with cBioPortal.}

\item{redcap_template}{A csv file with the REDCap data dictionary.}
}
\value{
A dataframe with data in cBioPortal format
}
\description{
Convert REDCap data to cBioPortal format
}
