% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{.source2CbioMap}
\alias{.source2CbioMap}
\title{Convert source-like Data to cBioPortal Format}
\usage{
.source2CbioMap(source_data, source_cbio_mapping)
}
\arguments{
\item{source_data}{A data frame containing the source data to be mapped}

\item{source_cbio_mapping}{A named list defining how fields from the source data should be mapped to cBioPortal fields}
}
\value{
A data frame compatible with the cBioPortal data structure.
}
\description{
This function takes a source data frame (e.g., extracted from REDCap), applies a predefined field mapping,
and returns a data frame formatted according to the structure expected by cBioPortal.
}
\examples{
\dontrun{
cbio_df <- .source2CbioMap(
  source_data = redcap_data,
  source_cbio_mapping = redcap_cbio_mapping
)
}
}
