% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{getStudyID}
\alias{getStudyID}
\title{Search study by REDCap URL and PID}
\usage{
getStudyID(studylink_url, access_token, redcap_url, pid)
}
\arguments{
\item{studylink_url}{The base URL of the StudyLink application}

\item{access_token}{A valid access token used for authentication with the StudyLink API}

\item{redcap_url}{The URL of the REDCap instance from which to retrieve the study}

\item{pid}{The project ID (PID) of the REDCap project whose study metadata should be retrieved}
}
\value{
The study ID corresponding to the specified REDCap study. Returns only one study or \code{NULL} if the request fails.
}
\description{
Searches and retrieves study metadata from a REDCap instance using the REDCap URL and project identifier (PID).
}
\examples{
\dontrun{
study_id <- getStudyID(
  studylink_url = "https://studylink.example.org",
  access_token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
  redcap_url = "https://redcap.example.org",
  pid = "1234"
)
}
}
