% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{getVialsAttributes}
\alias{getVialsAttributes}
\title{Get all vials with details and attributes by study id}
\usage{
getVialsAttributes(studylink_url, access_token, study_id, mapping_file = NULL)
}
\arguments{
\item{studylink_url}{The base URL of the StudyLink application}

\item{access_token}{A valid access token used for authentication with the StudyLink API}

\item{study_id}{The study ID}

\item{mapping_file}{A JSON file that provides field-level mapping rules to convert source data into a structure compatible with cBioPortal.}
}
\value{
A dataframe with data in cBioPortal format
}
\description{
Get all vials with details and attributes by study id
}
\examples{
\dontrun{
study_id <- getStudyID(
  studylink_url = "https://studylink.example.org",
  access_token = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
  redcap_url = "https://redcap.example.org",
  pid = "1234"
)
}
}
