% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REDCapToCbio-package.R
\docType{package}
\name{REDCapToCbio-package}
\alias{REDCapToCbio}
\alias{REDCapToCbio-package}
\title{REDCapToCbio: Formatting REDCap Data for cBioPortal Upload}
\description{
The package performs structured transformation of REDCap export data into the cBioPortal input format, ensuring compliance with the portal’s data schema for automated loading.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/biomeris/REDCapToCbio}
  \item Report bugs at \url{https://github.com/biomeris/REDCapToCbio/issues}
}

}
\author{
\strong{Maintainer}: Vittoria Ramella \email{vittoria.ramella@biomeris.it}

}
\keyword{internal}
