% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createClinicalDataFiles.R
\name{redcapToCbio}
\alias{redcapToCbio}
\title{Convert REDCap data to cBioPortal format}
\usage{
redcapToCbio(redcap_uri, token, mapping_file = NULL, redcap_template = NULL)
}
\arguments{
\item{redcap_uri}{The
\href{https://en.wikipedia.org/wiki/Uniform_Resource_Identifier}{uri}/url
of the REDCap server
typically formatted as "https://server.org/apps/redcap/api/".
Required.}

\item{token}{The user-specific string that serves as the password for a
project.  Required.}

\item{mapping_file}{A list that contains field-level mapping rules to convert source data into a structure compatible with cBioPortal.}

\item{redcap_template}{A csv file with the REDCap data dictionary.}
}
\value{
A dataframe with data in cBioPortal format
}
\description{
Convert REDCap data to cBioPortal format
}
\examples{
\dontrun{
redcap_data <- redcapToCbio(
  redcap_uri = "https://redcap-dev-2.ouhsc.edu/redcap/api/"
  token = "9A068C425B1341D69E83064A2D273A70"
)
}
}
