% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipStudyFolder.R
\name{zipStudyFolder}
\alias{zipStudyFolder}
\title{Zip a study folder}
\usage{
zipStudyFolder(zipfile, study_folder, output_path = getwd())
}
\arguments{
\item{zipfile}{Name of the resulting ZIP file (e.g., \code{"study.zip"}).}

\item{study_folder}{Path to the folder whose contents should be zipped.}

\item{output_path}{Directory where the ZIP file will be saved. Defaults to the current working directory.}
}
\description{
This function compresses the contents of a given study folder into a ZIP archive,
saving the output in a specified directory.
}
\examples{
\dontrun{
zipStudyFolder(
  zipfile = "study.zip",
  study_folder = "path/to/study",
  output_path = "path/to/output")
}
}
